+++ tools/sdk-ai-bots/azure-sdk-qa-bot-function/scripts/utils.ts
